move config for knex all to config.json, change to using postgres for development as well, add migrations for using jsonb instead of longtext, change underlying models to remove json stringifying and parsing steps
