+++ knexfile.js
@@ -1,13 +1,13 @@
 // Update with your config settings.
+const { loadEnvironmentVariables } = require('./src-backend/utils/handle-config')
+
+loadEnvironmentVariables()
 
 module.exports = {
 
   development: {
-    client: 'sqlite3',
-    connection: {
-      filename: './dev.sqlite3'
-    },
-    useNullAsDefault: true,
+    client: 'postgresql',
+    connection: process.env.PG_CONNECTION_STRING,
     migrations: {
       tableName: 'knex_migrations',
       directory: `${__dirname}/src-backend/db/migrations`
@@ -19,11 +19,7 @@ module.exports = {
 
   staging: {
     client: 'postgresql',
-    connection: {
-      database: 'my_db',
-      user: 'username',
-      password: 'password'
-    },
+    connection: process.env.PG_CONNECTION_STRING,
     pool: {
       min: 2,
       max: 10
@@ -39,11 +35,7 @@ module.exports = {
 
   production: {
     client: 'postgresql',
-    connection: {
-      database: 'my_db',
-      user: 'username',
-      password: 'password'
-    },
+    connection: process.env.PG_CONNECTION_STRING,
     pool: {
       min: 2,
       max: 10
